+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - order.customerをtransaction.object.customerから生成するように調整
 - プロダクトオファー承認時にポイント特典に識別子を自動発行
 - 注文作成&変更時にchevreへ連携
+- chevreでの注文検索に対応(USE_CHEVRE_SEARCH_ORDER設定)
 
 ### Deprecated